Make create-url-filter a separate file
